SDL_GetTicks replaced with a function with higher resolution in the Stopwatch class. The main loop of the InGame state now uses this new Stopwatch class. This new timer works with floating point values, thus allowing the TIMESTEP to be 16.666 instead of 16. The game now runs smoothly at a constant speed.

+++ premake/src/dyinganimation.cpp
@@ -23,7 +23,7 @@ void DyingAnimation::Animate()
 {
 	m_Sprite->Animate();
 
-	float t = m_Timer->GetCycleState() * 50.f;
+	float t = (float)(m_Timer->GetCycleState() * 50.f);
 	
 	glm::vec2 current_pos;
 	current_pos.x = m_StartPosition.x + (m_InitialVelocity.x * t * (float)cos(m_Angle * M_PI / 180.f));
